📜  README has end-user description
We've renamed the repo from `freenas_benchmarks` to
`cloud_storage_benchmarks` at Dmitriy Kalinin's suggestion.  He's right;
the repo consists of much more than mere FreeNAS benchmarks.

We've also added a brief description of how to use `jq` to parse
the output of the JSON, so the user has a sense of how to examine the
data.

We've also made the files of benchmarks that were run more than once
consistent.  We use a `-00` suffix to denote the results of the first
benchmark, `-01` for the second, etc.